+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v3.0.1] - 2023-05-16
+
+### Changed
+- More widely use import maps
+- Switch entirely over to unpkg
+
 ## [v3.0.0] - 2023-05-12
 
 ### Added